Clean up some APIs
