+++ drv/sprot-api/Cargo.toml
@@ -4,17 +4,18 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-crc = { workspace = true }
-derive_more = { workspace = true }
-gateway-messages = { workspace = true }
-hubpack = { workspace = true }
-idol-runtime = { workspace = true }
-memoffset = { workspace = true }
-num-traits = { workspace = true }
+crc.workspace = true
+derive_more.workspace = true
+gateway-messages.workspace = true
+hubpack.workspace = true
+idol-runtime.workspace = true
+memoffset.workspace = true
+num-traits.workspace = true
+serde.workspace = true
+sprockets-common.workspace = true
 static_assertions.workspace = true
-serde = { workspace = true }
-sprockets-common = { workspace = true }
-zerocopy = { workspace = true }
+tlvc.workspace = true
+zerocopy.workspace = true
 
 derive-idol-err = { path = "../../lib/derive-idol-err" }
 drv-caboose = { path = "../../drv/caboose" }
