+++ drv/sprot-api/src/lib.rs
@@ -9,10 +9,11 @@
 extern crate memoffset;
 
 mod error;
+use drv_caboose::CabooseError;
 use dumper_api::DumperError;
 pub use error::{
-    DumpOrSprotError, RawCabooseOrSprotError, SprocketsError, SprotError,
-    SprotProtocolError,
+    CabooseOrSprotError, DumpOrSprotError, RawCabooseOrSprotError,
+    SprocketsError, SprotError, SprotProtocolError,
 };
 
 use crc::{Crc, CRC_16_XMODEM};
@@ -524,6 +525,119 @@ pub struct SpIoStats {
 pub struct SprotIoStats {
     pub rot: RotIoStats,
     pub sp: SpIoStats,
+}
+
+impl SpRot {
+    pub fn read_caboose_value(
+        &self,
+        slot_id: SlotId,
+        key: [u8; 4],
+        buf: &mut [u8],
+    ) -> Result<u32, CabooseOrSprotError> {
+        let v =
+            RotCabooseReader::new(slot_id, &self).and_then(|r| r.get(key))?;
+        let len = v.len();
+        if len <= buf.len() {
+            self.read_caboose_region(v.start, slot_id, &mut buf[..len])?;
+            Ok(len as u32)
+        } else {
+            Err(CabooseOrSprotError::Sprot(SprotError::Protocol(
+                SprotProtocolError::BadMessageLength,
+            )))
+        }
+    }
+}
+
+#[derive(Copy, Clone)]
+struct RotCabooseReader<'a> {
+    sprot: &'a SpRot,
+    size: u32,
+    slot: SlotId,
+}
+
+impl<'a> RotCabooseReader<'a> {
+    fn new(
+        slot: SlotId,
+        sprot: &'a SpRot,
+    ) -> Result<Self, CabooseOrSprotError> {
+        let size = sprot.caboose_size(slot)?;
+        Ok(Self { size, slot, sprot })
+    }
+
+    pub fn get(
+        &self,
+        key: [u8; 4],
+    ) -> Result<core::ops::Range<u32>, CabooseOrSprotError> {
+        let mut reader = tlvc::TlvcReader::begin(self).map_err(|_| {
+            CabooseOrSprotError::Caboose(CabooseError::TlvcReaderBeginFailed)
+        })?;
+        loop {
+            match reader.next() {
+                Ok(Some(chunk)) => {
+                    if chunk.header().tag == key {
+                        let mut tmp = [0u8; 32];
+                        if chunk.check_body_checksum(&mut tmp).is_err() {
+                            return Err(CabooseOrSprotError::Caboose(
+                                CabooseError::BadChecksum,
+                            ));
+                        }
+                        // At this point, the reader is positioned **after** the data
+                        // from the target chunk.  We'll back up to the start of the
+                        // data slice.
+                        let (_reader, pos, _end) = reader.into_inner();
+
+                        let pos = pos as u32;
+                        let data_len = chunk.header().len.get();
+
+                        let data_start = pos
+                            - chunk.header().total_len_in_bytes() as u32
+                            + core::mem::size_of::<tlvc::ChunkHeader>() as u32;
+
+                        break Ok(data_start..(data_start + data_len));
+                    }
+                }
+                Err(e) => match e {
+                    tlvc::TlvcReadError::Truncated => {
+                        break Err(CabooseOrSprotError::Caboose(
+                            CabooseError::NoSuchTag,
+                        ))
+                    }
+                    tlvc::TlvcReadError::HeaderCorrupt { .. }
+                    | tlvc::TlvcReadError::BodyCorrupt { .. } => {
+                        break Err(CabooseOrSprotError::Caboose(
+                            CabooseError::BadChecksum,
+                        ))
+                    }
+                    tlvc::TlvcReadError::User(e) => break Err(e.into()),
+                },
+                Ok(None) => {
+                    break Err(CabooseOrSprotError::Caboose(
+                        CabooseError::NoSuchTag,
+                    ))
+                }
+            }
+        }
+    }
+}
+
+impl tlvc::TlvcRead for &RotCabooseReader<'_> {
+    type Error = RawCabooseOrSprotError;
+    fn extent(&self) -> Result<u64, tlvc::TlvcReadError<Self::Error>> {
+        Ok(self.size as u64)
+    }
+
+    fn read_exact(
+        &self,
+        offset: u64,
+        dest: &mut [u8],
+    ) -> Result<(), tlvc::TlvcReadError<Self::Error>> {
+        let offset = offset
+            .try_into()
+            .map_err(|_| tlvc::TlvcReadError::Truncated)?;
+        self.sprot
+            .read_caboose_region(offset, self.slot, dest)
+            .map_err(tlvc::TlvcReadError::User)
+    }
 }
 
 include!(concat!(env!("OUT_DIR"), "/client_stub.rs"));
