+++ drv/stm32h7-sprot-server/Cargo.toml
@@ -4,13 +4,13 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-cfg-if = { workspace = true }
-hubpack = { workspace = true }
-idol-runtime = { workspace = true }
-num-traits = { workspace = true }
-serde = { workspace = true }
-ssmarshal = { workspace = true }
-zerocopy = { workspace = true }
+cfg-if.workspace = true
+hubpack.workspace = true
+idol-runtime.workspace = true
+num-traits.workspace = true
+serde.workspace = true
+ssmarshal.workspace = true
+zerocopy.workspace = true
 
 mutable-statics = { path = "../../lib/mutable-statics" }
 drv-spi-api = { path = "../../drv/spi-api" }
